+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Run File Upload asset retention and stale pending upload cleanup through Craft garbage collection and dedicated console commands.
 
 - Add **Update on Submission Edit** to Entry element integrations. When **Update Entries** is enabled, this opt-in setting re-runs the integration when a submission is edited and re-submitted on the front end. ([#2591](https://github.com/verbb/formie/issues/2591), [#1970](https://github.com/verbb/formie/issues/1970))
+- Add **Use Submission User as Author** to Entry element integrations, assigning the collected submission user as the entry author when **Collect User** is enabled on the form. ([#1363](https://github.com/verbb/formie/issues/1363))
 
 ### Changed
 - Update phone number values sent to integrations to use `E164` format instead of international formatting. ([Discussion #2731](https://github.com/verbb/formie/discussions/2731))